Aca se sube el archivo de la programacion de la tarea2,se agrega la clase de notica para terminar el grupo 1, adicionalmente, se agregan las clases del material audiovisual, que es la principal y pelicula que es la clase derivada, todos con su respectivo archivo .hpp, finalmente se hacen ajustes minimos en el ambito de material de lectura y de libro

+++ tarea2/src/materialLectura.hpp
@@ -15,7 +15,7 @@ public:
     string genero;
     string estado;
     int cantidadHojas;
-    double precio;
+    int precio;
 
     MaterialLectura(const string& titulo, const string& grupo, const string& tipoMaterial,
                     const string& autor, const string& editorial, const string& genero,
